The one with the simplifications (#5)
This PR does a few things:

1. Swaps Hugo for Eleventy
Why: for the type of small site this project is intended for, I was starting to get bogged down in the opinionated structure of Hugo. Eleventy works like a drop-in replacement for jekyll, but it's written in JS and allows for a large variety of template languages right out of the box.

2. Cleans up the gulpfile
Why: again going for simplicity and small use cases. I'm probably going to strip out the webpacked javascript and image minification steps, or at least make them optional through build::webpack type sub-commands. I want this thing small and lean to start.